add optimizer registry (#2139)
* add optimizer registry

* move under core, add doc

* unexpose TORCH_OPTIMIZERS
