+++ mmdet/core/__init__.py
@@ -3,5 +3,6 @@ from .bbox import *  # noqa: F401, F403
 from .evaluation import *  # noqa: F401, F403
 from .fp16 import *  # noqa: F401, F403
 from .mask import *  # noqa: F401, F403
+from .optimizer import *  # noqa: F401, F403
 from .post_processing import *  # noqa: F401, F403
 from .utils import *  # noqa: F401, F403
