+++ mmdet/core/optimizer/__init__.py
@@ -1,0 +1,5 @@
+from .builder import build_optimizer
+from .copy_of_sgd import CopyOfSGD
+from .registry import OPTIMIZERS
+
+__all__ = ['OPTIMIZERS', 'build_optimizer', 'CopyOfSGD']
