+++ mmdet/core/optimizer/copy_of_sgd.py
@@ -1,0 +1,13 @@
+from torch.optim import SGD
+
+from .registry import OPTIMIZERS
+
+
+@OPTIMIZERS.register_module
+class CopyOfSGD(SGD):
+    """A clone of torch.optim.SGD.
+
+    A customized optimizer could be defined like CopyOfSGD.
+    You may derive from built-in optimizers in torch.optim,
+    or directly implement a new optimizer.
+    """
